Split tests before attaching NPU to CI
