+++ .github/workflows/test.yml
@@ -32,4 +32,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=100 --statistics
     - name: Test with pytest
       run: |
-        pytest
+        pytest --ignore=scripts
