+++ scripts/README.md
@@ -1,0 +1,14 @@
+# Model inference testing scripts
+
+Use these scripts to test Model before attaching NPU to the CI.
+
+## Tested Models
+- [x] ResNet34
+- [] ResNet50
+...
+
+## How to run
+Before running test, you need to install `furiosa-runtime`.
+You can install `furiosa-runtime` by running `pip install furiosa-runtime`.
+
+To run the test,`pytest ./scripts` in base directory of this repository.
