+++ scripts/assets/cat.jpg
